+++ CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
-- GraphQL queries will no longer show "__typename" in the request bodies ([Issue 20](https://github.com/rubrikinc/api-capture-chrome-extension/issues/20))
+- GraphQL queries will no longer show the "__typename" field in the request and response bodies ([Issue 20](https://github.com/rubrikinc/api-capture-chrome-extension/issues/20))
 
 ## v1.0.0